feat: add code coverage
Code coverage is now added - the coverage object is taken after each
spec end, from the iframe that run the test and is passed after each
spec end. All coverage objects are then merged and report is generated.
For this to work, the code has to have been previously instrumented with
Istanbul.