feat: report step progress to karma
Reporting progress to Karma for each step makes Karama less likely
to believe that the browser is disconnected when a test with many
steps is taking longer than usual. In practice, this makes Karma's
browserNoActivityTimeout limit the maximum allowed time for a step
(instead of a test) before considering the browser to be
disconnected.